Re-extract roleplay encounter screen components
