+++ apps/web/app/(app)/encounters/[id]/RoleplayEncounterScreens.tsx
@@ -9,7 +9,6 @@ import type {
   EncounterSession,
   RoleplayActionLogEntry,
   RoleplayCalculationPreview,
-  RoleplayDifficulty,
   RoleplayOpenEndedD20Roll,
   RoleplayParticipantDescription,
   Scenario,
@@ -54,15 +53,40 @@ import { buildCampaignWorkspaceHref } from "@/lib/campaigns/workspace";
 import {
   getPlayerFacingSkillBucket,
   getPlayerFacingSkillBucketDefinitions,
-  type PlayerFacingSkillBucketId,
 } from "@/lib/chargen/chargenBrowse";
 import { loadCanonicalContent } from "@/lib/content/loadCanonicalContent";
 
-interface RoleplayTopInfoProps {
-  campaignName?: string;
-  encounter: EncounterSession;
-  scenarioName?: string;
-}
+import { RoleplayRollBlock } from "./components/RoleplayRollBlock";
+import {
+  GmMessageSection,
+  ParticipantDescriptionsSection,
+  PlayerEncounterTopInfo,
+  RankedRollResultsSection,
+  RoleplayActionLogSection,
+  RoleplayTopInfo,
+  VisibilityGridSection,
+  formatShortDateTime,
+} from "./components/RoleplaySections";
+import { RoleplayRollCalculationPanel } from "./components/RoleplayCalculationPanel";
+import {
+  compactControlStyle,
+  compactInputStyle,
+  compactSkillInputStyle,
+  panelStyle,
+  playerReadOnlyPanelStyle,
+  playerRollSkillColumnsStyle,
+  rollBlockShellStyle,
+  rollControlRowStyle,
+  rollControlsStackStyle,
+  rollControlsStyle,
+  rollEditorStyle,
+  rollFieldRowStyle,
+} from "./components/roleplayStyles";
+import type {
+  PlayerLocalRollDraft,
+  RoleplayRollDraft,
+  SkillOption,
+} from "./components/roleplayRollTypes";
 
 interface GmRoleplayingEncounterScreenProps {
   campaignId?: string;
@@ -81,210 +105,6 @@ interface PlayerRoleplayingEncounterScreenProps {
   embedded?: boolean;
   encounterId: string;
   scenarioId: string;
-}
-
-interface SkillOption {
-  categoryId?: PlayerFacingSkillBucketId;
-  categoryLabel?: string;
-  id: string;
-  label: string;
-  profile?: ParticipantSkillRollProfile;
-  value?: number;
-  warning?: string;
-}
-
-interface RoleplayRollDraft {
-  actorRoll?: RoleplayOpenEndedD20Roll;
-  actorSupportRoll?: RoleplayOpenEndedD20Roll;
-  difficulty: "none" | RoleplayDifficulty;
-  id: string;
-  otherModInput: string;
-  otherModTouched: boolean;
-  opponentBlockOpen: boolean;
-  opponentOtherModInput: string;
-  opponentOtherModTouched: boolean;
-  opponentParticipantId: string;
-  opponentRoll?: RoleplayOpenEndedD20Roll;
-  opponentSupportRoll?: RoleplayOpenEndedD20Roll;
-  opponentSilent: boolean;
-  opponentSkillCategoryId: "all" | PlayerFacingSkillBucketId;
-  opponentSkillId: string;
-  opponentSupportSkillCategoryId: "all" | PlayerFacingSkillBucketId;
-  opponentSupportSkillId: string;
-  opponentUseDbMod: boolean;
-  opponentUseGenMod: boolean;
-  opponentUseObSkillMod: boolean;
-  participantId: string;
-  rollSetId: string;
-  silent: boolean;
-  skillCategoryId: "all" | PlayerFacingSkillBucketId;
-  skillId: string;
-  supportSkillCategoryId: "all" | PlayerFacingSkillBucketId;
-  supportSkillId: string;
-  useDbMod: boolean;
-  useGenMod: boolean;
-  useObSkillMod: boolean;
-}
-
-interface PlayerLocalRollDraft {
-  difficulty: "none" | RoleplayDifficulty;
-  otherModInput: string;
-  otherModTouched: boolean;
-  opponentParticipantId: string;
-  roll?: RoleplayOpenEndedD20Roll;
-  supportRoll?: RoleplayOpenEndedD20Roll;
-  skillCategoryId: "all" | PlayerFacingSkillBucketId;
-  skillId: string;
-  supportSkillCategoryId: "all" | PlayerFacingSkillBucketId;
-  supportSkillId: string;
-  useDbMod: boolean;
-  useGenMod: boolean;
-  useObSkillMod: boolean;
-}
-
-const panelStyle = {
-  border: "1px solid #d9ddd8",
-  borderRadius: 12,
-  display: "grid",
-  gap: "0.75rem",
-  padding: "1rem",
-} as const;
-
-const compactControlStyle = {
-  display: "grid",
-  gap: "0.2rem",
-} as const;
-
-const compactInputStyle = {
-  maxWidth: "8.75rem",
-  minHeight: "1.9rem",
-  width: "8.75rem",
-} as const;
-
-const compactSkillInputStyle = {
-  ...compactInputStyle,
-  maxWidth: "10.5rem",
-  width: "10.5rem",
-} as const;
-
-const rollBlockShellStyle = {
-  border: "1px solid #eee8dc",
-  borderRadius: 10,
-  display: "grid",
-  gap: "0.6rem",
-  padding: "0.75rem",
-  overflow: "hidden",
-} as const;
-
-const rollEditorStyle = {
-  alignItems: "start",
-  display: "grid",
-  gap: "0.85rem",
-  gridTemplateColumns: "minmax(0, 1fr) minmax(22rem, 1fr)",
-  minWidth: 0,
-} as const;
-
-const rollControlsStackStyle = {
-  display: "grid",
-  gap: "0.9rem",
-  minWidth: 0,
-} as const;
-
-const rollControlsStyle = {
-  display: "grid",
-  gap: "0.65rem",
-  minWidth: 0,
-} as const;
-
-const opponentControlsStyle = {
-  ...rollControlsStyle,
-  borderTop: "1px solid #eee8dc",
-  paddingTop: "0.9rem",
-} as const;
-
-const rollControlRowStyle = {
-  alignItems: "center",
-  display: "flex",
-  flexWrap: "wrap",
-  gap: "0.45rem",
-} as const;
-
-const rollFieldRowStyle = {
-  ...rollControlRowStyle,
-  alignItems: "end",
-} as const;
-
-const rollSkillGridStyle = {
-  display: "grid",
-  gap: "0.45rem 0.65rem",
-  gridTemplateColumns: "repeat(2, minmax(0, max-content))",
-} as const;
-
-const rollPreviewStyle = {
-  background: "#fbfaf7",
-  border: "1px solid #eee8dc",
-  borderRadius: 8,
-  color: "#5e5a50",
-  display: "grid",
-  gap: "0.35rem",
-  gridTemplateRows: "auto",
-  alignSelf: "stretch",
-  boxSizing: "border-box",
-  justifySelf: "stretch",
-  minHeight: "10.5rem",
-  minWidth: 0,
-  padding: "0.65rem",
-  width: "100%",
-} as const;
-
-const calculationLineStyle = {
-  display: "block",
-  overflowX: "auto",
-  whiteSpace: "nowrap",
-} as const;
-
-const playerReadOnlyPanelStyle = {
-  background: "#fbfaf7",
-  border: "1px solid #eee8dc",
-  borderRadius: 8,
-  padding: "0.75rem",
-  whiteSpace: "pre-wrap",
-} as const;
-
-const playerMetadataTagStyle = {
-  flex: "0 1 auto",
-  fontSize: "0.95rem",
-  fontWeight: 400,
-  minWidth: 0,
-  overflow: "hidden",
-  textOverflow: "ellipsis",
-  whiteSpace: "nowrap",
-} as const;
-
-const playerRollSkillColumnsStyle = {
-  alignItems: "end",
-  display: "grid",
-  gap: "0.45rem",
-  gridTemplateColumns: "minmax(8rem, 9rem) minmax(10rem, 12rem)",
-} as const;
-
-function formatDifficulty(value: RoleplayDifficulty): string {
-  return roleplayDifficultyOptions.find((option) => option.id === value)?.label ?? value;
-}
-
-function formatShortDateTime(value: string): string {
-  const date = new Date(value);
-
-  if (Number.isNaN(date.getTime())) {
-    return value;
-  }
-
-  const day = String(date.getDate()).padStart(2, "0");
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const hours = String(date.getHours()).padStart(2, "0");
-  const minutes = String(date.getMinutes()).padStart(2, "0");
-
-  return `${day}.${month} ${hours}:${minutes}`;
 }
 
 function isRecord(value: unknown): value is Record<string, unknown> {
@@ -492,203 +312,6 @@ function buildRoleplayCalculationPreview(
     modifierPipeline,
     otherMod,
   });
-}
-
-function RollCalculationPreview({
-  cleanPendingText = false,
-  label,
-  pendingLabels,
-  preview,
-  showPendingLabels = true,
-}: {
-  cleanPendingText?: boolean;
-  label: string;
-  pendingLabels?: string[];
-  preview?: ReturnType<typeof buildRoleplayCalculationPreview>;
-  showPendingLabels?: boolean;
-}) {
-  const formulaText = preview
-    ? cleanPendingText
-      ? preview.formulaText.replace(" = pending", " = —")
-      : preview.formulaText
-    : "—";
-
-  return (
-    <div style={calculationLineStyle}>
-      <strong>{label}:</strong> {formulaText}
-      {showPendingLabels && preview && pendingLabels && pendingLabels.length > 0 ? (
-        <span> · Pending: {pendingLabels.join(", ")}</span>
-      ) : null}
-    </div>
-  );
-}
-
-function getPreviewResultLabel(input: {
-  comparison?: string;
-  difficulty?: RoleplayDifficulty;
-  preview?: ReturnType<typeof buildRoleplayCalculationPreview>;
-}): string {
-  if (input.comparison) {
-    return input.comparison;
-  }
-
-  const preview = input.preview;
-
-  if (!preview) {
-    return "—";
-  }
-
-  if (preview.fumble) {
-    return preview.resultText ?? "FUMBLE";
-  }
-
-  if (preview.autoSuccess) {
-    return preview.resultText ?? "Automatic success";
-  }
-
-  if (preview.resultText) {
-    return preview.resultText;
-  }
-
-  return preview.achievedSuccessLevel?.label ?? "—";
-}
-
-function RoleplayRollResultLine({
-  comparison,
-  difficulty,
-  label = "Result",
-  preview,
-}: {
-  comparison?: string;
-  difficulty?: RoleplayDifficulty;
-  label?: string;
-  preview?: ReturnType<typeof buildRoleplayCalculationPreview>;
-}) {
-  return <div><strong>{label}:</strong> {getPreviewResultLabel({ comparison, difficulty, preview })}</div>;
-}
-
-function RoleplayRollCalculationPanel({
-  actorDifficulty,
-  actorLabel = "Actor",
-  actorMainPreview,
-  actorPendingLabels,
-  actorSupportPreview,
-  cleanPendingText = false,
-  comparison,
-  opponentMainPreview,
-  opponentLabel = "Opponent",
-  opponentOpen,
-  opponentPendingLabels,
-  showPendingLabels = true,
-  opponentSupportPreview,
-}: {
-  actorDifficulty?: RoleplayDifficulty;
-  actorLabel?: string;
-  actorMainPreview?: ReturnType<typeof buildRoleplayCalculationPreview>;
-  actorPendingLabels?: string[];
-  actorSupportPreview?: ReturnType<typeof buildRoleplayCalculationPreview>;
-  cleanPendingText?: boolean;
-  comparison?: string;
-  opponentMainPreview?: ReturnType<typeof buildRoleplayCalculationPreview>;
-  opponentLabel?: string;
-  opponentOpen: boolean;
-  opponentPendingLabels?: string[];
-  showPendingLabels?: boolean;
-  opponentSupportPreview?: ReturnType<typeof buildRoleplayCalculationPreview>;
-}) {
-  return (
-    <div style={rollPreviewStyle}>
-      <strong>Calculation</strong>
-      <div style={{ display: "grid", gap: "0.25rem" }}>
-        <strong>{actorLabel}</strong>
-        <RollCalculationPreview cleanPendingText={cleanPendingText} label="Support" preview={actorSupportPreview} />
-        <RollCalculationPreview
-          cleanPendingText={cleanPendingText}
-          label="Main"
-          pendingLabels={actorPendingLabels}
-          preview={actorMainPreview}
-          showPendingLabels={showPendingLabels}
-        />
-        <RoleplayRollResultLine difficulty={actorDifficulty} preview={actorMainPreview} />
-      </div>
-      {opponentOpen ? (
-        <div style={{ borderTop: "1px solid #eee8dc", display: "grid", gap: "0.25rem", paddingTop: "0.45rem" }}>
-          <strong>{opponentLabel}</strong>
-          <RollCalculationPreview cleanPendingText={cleanPendingText} label="Support" preview={opponentSupportPreview} />
-          <RollCalculationPreview
-            cleanPendingText={cleanPendingText}
-            label="Main"
-            pendingLabels={opponentPendingLabels}
-            preview={opponentMainPreview}
-            showPendingLabels={showPendingLabels}
-          />
-          <RoleplayRollResultLine preview={opponentMainPreview} />
-        </div>
-      ) : null}
-      {comparison ? (
-        <div style={{ borderTop: "1px solid #eee8dc", paddingTop: "0.45rem" }}>
-          <RoleplayRollResultLine comparison={comparison} label="Comparison" />
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function getParticipantDescription(input: {
-  fallbackName: string;
-  participantId: string;
-  state: ReturnType<typeof normalizeRoleplayState>;
-}): RoleplayParticipantDescription {
-  const existing = input.state.participantDescriptions[input.participantId];
-
-  return {
-    detailedDescription: existing?.detailedDescription ?? "",
-    name: existing?.name ?? input.fallbackName,
-    shortDescription: existing?.shortDescription ?? "",
-  };
-}
-
-function RoleplayTopInfo({ campaignName, encounter, scenarioName }: RoleplayTopInfoProps) {
-  return (
-    <section style={panelStyle}>
-      <h1 style={{ margin: 0 }}>{encounter.title}</h1>
-      <div style={{ display: "flex", flexWrap: "wrap", gap: "0.75rem" }}>
-        <span>Type: Roleplaying</span>
-        <span>Status: {encounter.status}</span>
-        {encounter.timelineLabel ? <span>Timeline: {encounter.timelineLabel}</span> : null}
-        {scenarioName ? <span>Scenario: {scenarioName}</span> : null}
-        {campaignName ? <span>Campaign: {campaignName}</span> : null}
-      </div>
-      {encounter.description ? <div>{encounter.description}</div> : null}
-    </section>
-  );
-}
-
-function PlayerEncounterTopInfo({ campaignName, encounter, scenarioName }: RoleplayTopInfoProps) {
-  return (
-    <section aria-label="Player encounter summary" style={panelStyle}>
-      <div style={{ alignItems: "baseline", display: "flex", gap: "0.5rem", minWidth: 0, overflow: "hidden", whiteSpace: "nowrap" }}>
-        <h1 style={{ flex: "0 0 auto", fontSize: "1.35rem", lineHeight: 1.2, margin: 0 }}>
-          {encounter.title}
-        </h1>
-        <span aria-hidden="true" style={playerMetadataTagStyle}>·</span>
-        <span style={playerMetadataTagStyle}>{encounter.kind === "roleplay" ? "Roleplaying" : "Combat"}</span>
-        {scenarioName ? (
-          <>
-            <span aria-hidden="true" style={playerMetadataTagStyle}>·</span>
-            <span style={playerMetadataTagStyle}>Scenario: {scenarioName}</span>
-          </>
-        ) : null}
-        {campaignName ? (
-          <>
-            <span aria-hidden="true" style={playerMetadataTagStyle}>·</span>
-            <span style={playerMetadataTagStyle}>Campaign: {campaignName}</span>
-          </>
-        ) : null}
-      </div>
-      {encounter.description ? <div style={playerReadOnlyPanelStyle}>{encounter.description}</div> : null}
-    </section>
-  );
 }
 
 function buildPlayerRollResult(input?: {
@@ -992,6 +615,36 @@ export function GmRoleplayingEncounterScreen({
   function updateRollDraft(draftId: string, patch: Partial<RoleplayRollDraft>) {
     setRollDrafts((currentDrafts) =>
       currentDrafts.map((draft) => (draft.id === draftId ? { ...draft, ...patch } : draft))
+    );
+  }
+
+  function handleActorParticipantChange(draftId: string, participantId: string) {
+    setRollDrafts((currentDrafts) =>
+      currentDrafts.map((draft) => {
+        if (draft.id !== draftId) {
+          return draft;
+        }
+
+        const participant = roster.find((row) => row.id === participantId);
+        const nextSkillOptions = readSystemSkillOptions({
+          content,
+          encounterParticipant: participant,
+          scenarioParticipants,
+        });
+        const nextSkillId = nextSkillOptions[0]?.id ?? "";
+
+        return {
+          ...draft,
+          otherModInput: applyUnknownSkillDefaultOtherMod({
+            currentValue: draft.otherModInput,
+            selectedSkill: nextSkillOptions.find((skill) => skill.id === nextSkillId),
+            touched: draft.otherModTouched,
+          }),
+          participantId,
+          skillCategoryId: "all",
+          skillId: nextSkillId,
+        };
+      })
     );
   }
 
@@ -1648,100 +1301,24 @@ export function GmRoleplayingEncounterScreen({
         scenarioName={scenario?.name}
       />
 
-      <section style={panelStyle}>
-        <h2 style={{ margin: 0 }}>GM message</h2>
-        <textarea
-          onChange={(event) => setGmMessageDraft(event.target.value)}
-          placeholder="Message intended for player roleplaying screens."
-          rows={4}
-          value={gmMessageDraft}
-        />
-        <button onClick={() => void handleSaveGmMessage()} type="button">
-          Save GM message
-        </button>
-      </section>
+      <GmMessageSection
+        gmMessageDraft={gmMessageDraft}
+        onChange={setGmMessageDraft}
+        onSave={handleSaveGmMessage}
+      />
 
-      <section style={panelStyle}>
-        <h2 style={{ margin: 0 }}>Visibility grid</h2>
-        {roster.length > 0 ? (
-          <div style={{ overflowX: "auto" }}>
-            <table style={{ borderCollapse: "collapse", minWidth: 760, width: "100%" }}>
-              <thead>
-                <tr style={{ borderBottom: "1px solid #d9ddd8", textAlign: "left" }}>
-                  <th style={{ padding: "0.5rem 0.75rem 0.5rem 0" }}>Viewer</th>
-                  <th style={{ padding: "0.5rem 0.75rem" }}>Row action</th>
-                  {roster.map((target) => (
-                    <th key={target.id} style={{ padding: "0.5rem 0.75rem", textAlign: "center" }}>
-                      {target.label}
-                    </th>
-                  ))}
-                </tr>
-              </thead>
-              <tbody>
-                {roster.map((viewer) => (
-                  <tr key={viewer.id} style={{ borderBottom: "1px solid #eee8dc" }}>
-                    <td style={{ padding: "0.5rem 0.75rem 0.5rem 0" }}>{viewer.label}</td>
-                    <td style={{ padding: "0.5rem 0.75rem" }}>
-                      <button onClick={() => void handleSelectAllVisibility(viewer.id)} type="button">
-                        Select all
-                      </button>
-                    </td>
-                    {roster.map((target) => (
-                      <td key={target.id} style={{ padding: "0.5rem 0.75rem", textAlign: "center" }}>
-                        <input
-                          aria-label={`${viewer.label} can see ${target.label}`}
-                          checked={
-                            viewer.id === target.id ||
-                            Boolean(roleplayState.visibility[viewer.id]?.[target.id])
-                          }
-                          disabled={viewer.id === target.id}
-                          onChange={(event) =>
-                            void handleVisibilityChange({
-                              targetParticipantId: target.id,
-                              viewerParticipantId: viewer.id,
-                              visible: event.target.checked,
-                            })
-                          }
-                          type="checkbox"
-                        />
-                      </td>
-                    ))}
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        ) : (
-          <div>Add participants to this encounter from the Scenario screen before setting visibility.</div>
-        )}
-      </section>
+      <VisibilityGridSection
+        onSelectAllVisibility={handleSelectAllVisibility}
+        onVisibilityChange={handleVisibilityChange}
+        roster={roster}
+        state={roleplayState}
+      />
 
-      <section style={panelStyle}>
-        <h2 style={{ margin: 0 }}>Roleplay roster descriptions</h2>
-        {roster.length > 0 ? (
-          <div style={{ display: "grid", gap: "0.75rem", maxHeight: "36rem", overflow: "auto" }}>
-            {roster.map((participant) => (
-              <RoleplayDescriptionRow
-                description={getParticipantDescription({
-                  fallbackName: participant.label,
-                  participantId: participant.id,
-                  state: roleplayState,
-                })}
-                key={participant.id}
-                onSave={(description) =>
-                  void handleDescriptionSave({
-                    description,
-                    participantId: participant.id,
-                  })
-                }
-                participant={participant}
-              />
-            ))}
-          </div>
-        ) : (
-          <div>No participants are assigned to this roleplaying encounter yet.</div>
-        )}
-      </section>
+      <ParticipantDescriptionsSection
+        onSave={handleDescriptionSave}
+        roster={roster}
+        state={roleplayState}
+      />
 
       <section style={panelStyle}>
         <div style={{ alignItems: "center", display: "flex", justifyContent: "space-between", gap: "0.75rem" }}>
@@ -1753,19 +1330,6 @@ export function GmRoleplayingEncounterScreen({
         <div style={{ display: "grid", gap: "0.75rem" }}>
           {rollDrafts.map((draft, index) => {
             const context = getRollDraftContext(draft);
-            const categoryOptions = getPlayerFacingSkillBucketDefinitions().filter((category) =>
-              context.allSkillOptions.some((skill) => skill.categoryId === category.id)
-            );
-            const supportCategoryOptions = getPlayerFacingSkillBucketDefinitions().filter((category) =>
-              context.allSkillOptions.some((skill) => skill.categoryId === category.id)
-            );
-            const opponentCategoryOptions = getPlayerFacingSkillBucketDefinitions().filter((category) =>
-              context.allOpponentSkillOptions.some((skill) => skill.categoryId === category.id)
-            );
-
-            const opponentSupportCategoryOptions = getPlayerFacingSkillBucketDefinitions().filter((category) =>
-              context.allOpponentSkillOptions.some((skill) => skill.categoryId === category.id)
-            );
             const comparison =
               context.preview && context.opponentPreview
                 ? compareRoleplayOpposedRolls({
@@ -1775,580 +1339,21 @@ export function GmRoleplayingEncounterScreen({
                     opponentPreview: context.opponentPreview,
                   }).summary
                 : undefined;
-            const actorLocked = Boolean(draft.actorRoll || context.actorExternalResult);
-            const opponentLocked = Boolean(draft.opponentRoll || context.opponentExternalResult);
-            const actorStatus = context.actorExternalResult
-              ? "Result received"
-              : draft.actorRoll
-                ? "Rolled"
-                : context.matchingPendingRoll
-                  ? "Assigned · Pending player roll"
-                  : undefined;
-            const opponentStatus = context.opponentExternalResult
-              ? "Result received"
-              : draft.opponentRoll
-                ? "Rolled"
-                : context.matchingOpponentPendingRoll
-                  ? "Assigned · Pending player roll"
-                  : undefined;
 
             return (
-              <div key={draft.id} style={rollBlockShellStyle}>
-                <section style={rollEditorStyle}>
-                  <div style={rollControlsStackStyle}>
-                    <section style={rollControlsStyle}>
-                    <div style={{ alignItems: "center", display: "flex", flexWrap: "wrap", gap: "0.55rem" }}>
-                      <strong>Roll {index + 1}</strong>
-                      <select
-                        aria-label={`Roleplay roll ${index + 1} participant`}
-                        disabled={actorLocked}
-                        onChange={(event) => {
-                          const participant = roster.find((row) => row.id === event.target.value);
-                          const nextSkillId = readSystemSkillOptions({
-                            content,
-                            encounterParticipant: participant,
-                            scenarioParticipants,
-                          })[0]?.id ?? "";
-                          updateRollDraft(draft.id, {
-                            otherModInput: applyUnknownSkillDefaultOtherMod({
-                              currentValue: draft.otherModInput,
-                              selectedSkill: readSystemSkillOptions({
-                                content,
-                                encounterParticipant: participant,
-                                scenarioParticipants,
-                              }).find((skill) => skill.id === nextSkillId),
-                              touched: draft.otherModTouched,
-                            }),
-                            participantId: event.target.value,
-                            skillCategoryId: "all",
-                            skillId: nextSkillId,
-                          });
-                        }}
-                        style={compactInputStyle}
-                        value={context.participant?.id ?? ""}
-                      >
-                        {roster.map((participant) => (
-                          <option key={participant.id} value={participant.id}>
-                            {participant.label}
-                          </option>
-                        ))}
-                      </select>
-                    </div>
-                    <div style={rollControlRowStyle}>
-                      <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                        <input
-                          checked={draft.silent}
-                          disabled={actorLocked}
-                          onChange={(event) => updateRollDraft(draft.id, { silent: event.target.checked })}
-                          type="checkbox"
-                        />
-                        Silent
-                      </label>
-                      <span>Use:</span>
-                      <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                        <input
-                          checked={draft.useGenMod}
-                          disabled={actorLocked}
-                          onChange={(event) => updateRollDraft(draft.id, { useGenMod: event.target.checked })}
-                          type="checkbox"
-                        />
-                        Gen
-                      </label>
-                      <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                        <input
-                          checked={draft.useObSkillMod}
-                          disabled={actorLocked}
-                          onChange={(event) => updateRollDraft(draft.id, { useObSkillMod: event.target.checked })}
-                          type="checkbox"
-                        />
-                        OB/Skill
-                      </label>
-                      <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                        <input
-                          checked={draft.useDbMod}
-                          disabled={actorLocked}
-                          onChange={(event) => updateRollDraft(draft.id, { useDbMod: event.target.checked })}
-                          type="checkbox"
-                        />
-                        DB
-                      </label>
-                      <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                        Other
-                        <input
-                          aria-label={`Roleplay roll ${index + 1} Other mod`}
-                          disabled={actorLocked}
-                          onChange={(event) =>
-                            updateRollDraft(draft.id, {
-                              otherModInput: event.target.value,
-                              otherModTouched: true,
-                            })
-                          }
-                          step={1}
-                          style={{ ...compactInputStyle, width: "4.5rem" }}
-                          type="number"
-                          value={context.actorOtherModInput}
-                        />
-                      </label>
-                    </div>
-                    <div style={rollSkillGridStyle}>
-                      <label style={compactControlStyle}>
-                        <span>Category</span>
-                        <select
-                          aria-label={`Roleplay roll ${index + 1} skill category`}
-                          disabled={actorLocked}
-                          onChange={(event) => {
-                            const nextCategoryId = event.target.value as RoleplayRollDraft["skillCategoryId"];
-                            const nextSkillOptions =
-                              nextCategoryId === "all"
-                                ? context.allSkillOptions
-                                : context.allSkillOptions.filter((skill) => skill.categoryId === nextCategoryId);
-
-                            updateRollDraft(draft.id, {
-                              otherModInput: applyUnknownSkillDefaultOtherMod({
-                                currentValue: draft.otherModInput,
-                                selectedSkill: nextSkillOptions[0],
-                                touched: draft.otherModTouched,
-                              }),
-                              skillCategoryId: nextCategoryId,
-                              skillId: nextSkillOptions[0]?.id ?? "",
-                            });
-                          }}
-                          style={compactInputStyle}
-                          value={draft.skillCategoryId}
-                        >
-                          <option value="all">All categories</option>
-                          {categoryOptions.map((category) => (
-                            <option key={category.id} value={category.id}>
-                              {category.label}
-                            </option>
-                          ))}
-                        </select>
-                      </label>
-                      <label style={compactControlStyle}>
-                        <span>Skill</span>
-                        <select
-                          aria-label={`Roleplay roll ${index + 1} skill`}
-                          disabled={actorLocked || context.skillOptions.length === 0}
-                          onChange={(event) =>
-                            updateRollDraft(draft.id, {
-                              otherModInput: applyUnknownSkillDefaultOtherMod({
-                                currentValue: draft.otherModInput,
-                                selectedSkill: context.skillOptions.find((skill) => skill.id === event.target.value),
-                                touched: draft.otherModTouched,
-                              }),
-                              skillId: event.target.value,
-                            })
-                          }
-                          style={compactSkillInputStyle}
-                          value={context.selectedSkill?.id ?? ""}
-                        >
-                          {context.skillOptions.length > 0 ? (
-                            context.skillOptions.map((skill) => (
-                              <option key={skill.id} value={skill.id}>
-                                {skill.label} ({skill.value ?? 0})
-                              </option>
-                            ))
-                          ) : (
-                            <option value="">No skills available</option>
-                          )}
-                        </select>
-                      </label>
-                      <label style={compactControlStyle}>
-                        <span>Support category</span>
-                        <select
-                          aria-label={`Roleplay roll ${index + 1} support skill category`}
-                          disabled={actorLocked}
-                          onChange={(event) => {
-                            const nextCategoryId = event.target.value as RoleplayRollDraft["supportSkillCategoryId"];
-                            updateRollDraft(draft.id, {
-                              supportSkillCategoryId: nextCategoryId,
-                              supportSkillId: "",
-                            });
-                          }}
-                          style={compactInputStyle}
-                          value={draft.supportSkillCategoryId}
-                        >
-                          <option value="all">All categories</option>
-                          {supportCategoryOptions.map((category) => (
-                            <option key={category.id} value={category.id}>
-                              {category.label}
-                            </option>
-                          ))}
-                        </select>
-                      </label>
-                      <label style={compactControlStyle}>
-                        <span>Support</span>
-                        <select
-                          aria-label={`Roleplay roll ${index + 1} support skill`}
-                          disabled={actorLocked}
-                          onChange={(event) => updateRollDraft(draft.id, { supportSkillId: event.target.value })}
-                          style={compactSkillInputStyle}
-                          value={context.selectedSupportSkill?.id ?? ""}
-                        >
-                          <option value="">No support skill</option>
-                          {context.supportSkillOptions.map((skill) => (
-                            <option key={skill.id} value={skill.id}>
-                              {skill.label} ({skill.value ?? 0})
-                            </option>
-                          ))}
-                        </select>
-                      </label>
-                    </div>
-                    <div style={rollFieldRowStyle}>
-                      <label style={compactControlStyle}>
-                        <span>Level</span>
-                        <select
-                          aria-label={`Roleplay roll ${index + 1} difficulty`}
-                          disabled={actorLocked}
-                          onChange={(event) => {
-                            const nextDifficulty = event.target.value as RoleplayRollDraft["difficulty"];
-                            updateRollDraft(draft.id, {
-                              difficulty: nextDifficulty,
-                              opponentBlockOpen: nextDifficulty === "none" ? draft.opponentBlockOpen : false,
-                              opponentOtherModInput: nextDifficulty === "none" ? draft.opponentOtherModInput : "0",
-                              opponentOtherModTouched: nextDifficulty === "none" ? draft.opponentOtherModTouched : false,
-                              opponentParticipantId: nextDifficulty === "none" ? draft.opponentParticipantId : "",
-                              opponentRoll: undefined,
-                              opponentSkillCategoryId: "all",
-                              opponentSkillId: "",
-                              opponentSupportSkillCategoryId: "all",
-                              opponentSupportSkillId: "",
-                            });
-                          }}
-                          style={compactInputStyle}
-                          value={draft.difficulty}
-                        >
-                          <option value="none">No level</option>
-                          {roleplayDifficultyOptions.map((option) => (
-                            <option key={option.id} value={option.id}>
-                              {option.label}
-                            </option>
-                          ))}
-                        </select>
-                      </label>
-                      <label style={compactControlStyle}>
-                        <span>Opponent</span>
-                        <select
-                          aria-label={`Roleplay roll ${index + 1} opponent`}
-                          disabled={actorLocked}
-                          onChange={(event) => {
-                            updateRollDraft(draft.id, {
-                              difficulty: "none",
-                              opponentBlockOpen: Boolean(event.target.value) && draft.opponentBlockOpen,
-                              opponentOtherModInput: "0",
-                              opponentOtherModTouched: false,
-                              opponentParticipantId: event.target.value,
-                              opponentRoll: undefined,
-                              opponentSkillCategoryId: "all",
-                              opponentSkillId: "",
-                              opponentSupportSkillCategoryId: "all",
-                              opponentSupportSkillId: "",
-                            });
-                          }}
-                          style={compactInputStyle}
-                          value={context.opponent?.id ?? ""}
-                        >
-                          <option value="">No opponent</option>
-                          {roster.map((participant) => (
-                            <option key={participant.id} value={participant.id}>
-                              {participant.label}
-                            </option>
-                          ))}
-                        </select>
-                      </label>
-                    </div>
-                    {[context.selectedSkill, context.selectedSupportSkill]
-                      .filter((skill): skill is SkillOption => Boolean(skill?.warning))
-                      .map((skill) => (
-                        <div key={skill.id} style={{ color: "#8a5a00", fontSize: "0.85rem" }}>
-                          {skill.label}: {skill.warning}
-                        </div>
-                      ))}
-                    <div style={{ display: "flex", flexWrap: "wrap", gap: "0.5rem" }}>
-                      {actorStatus ? <span style={{ color: "#5e5a50" }}>{actorStatus}</span> : null}
-                      {!context.opponent || draft.opponentBlockOpen ? (
-                        <button
-                          disabled={actorLocked || !context.participant || !context.selectedSkill}
-                          onClick={() => void handleAssignSkillRoll(draft, "actor")}
-                          type="button"
-                        >
-                          Assign
-                        </button>
-                      ) : null}
-                      {!context.opponent || draft.opponentBlockOpen ? (
-                        <button
-                          disabled={actorLocked || !context.participant || !context.selectedSkill}
-                          onClick={() => void handleGmRoll(draft, "actor")}
-                          type="button"
-                        >
-                          GM Roll
-                        </button>
-                      ) : !draft.opponentBlockOpen ? (
-                        <button
-                          disabled={actorLocked || !context.opponent}
-                          onClick={() => {
-                            updateRollDraft(draft.id, {
-                              opponentBlockOpen: true,
-                              opponentSkillId: "",
-                              opponentRoll: undefined,
-                            });
-                          }}
-                          type="button"
-                        >
-                          Open opponent block
-                        </button>
-                      ) : null}
-                    </div>
-                    {context.isOpposed ? (
-                      <div style={{ display: "flex", justifyContent: "center" }}>
-                        <button
-                          disabled={
-                            actorLocked ||
-                            opponentLocked ||
-                            !context.participant ||
-                            !context.selectedSkill ||
-                            !context.selectedOpponentSkill
-                          }
-                          onClick={() => void handleGmRoll(draft, "both")}
-                          type="button"
-                        >
-                          GM Roll both
-                        </button>
-                      </div>
-                    ) : null}
-                    </section>
-                    {context.opponent && draft.opponentBlockOpen ? (
-                    <section style={opponentControlsStyle}>
-                      <div style={{ alignItems: "center", display: "flex", flexWrap: "wrap", gap: "0.55rem" }}>
-                        <strong>Opponent</strong>
-                        <select
-                          aria-label={`Roleplay roll ${index + 1} opponent block participant`}
-                          disabled={opponentLocked}
-                          onChange={(event) => {
-                            updateRollDraft(draft.id, {
-                              opponentParticipantId: event.target.value,
-                              opponentRoll: undefined,
-                              opponentOtherModInput: "0",
-                              opponentOtherModTouched: false,
-                              opponentSkillCategoryId: "all",
-                              opponentSkillId: "",
-                              opponentSupportSkillCategoryId: "all",
-                              opponentSupportSkillId: "",
-                            });
-                          }}
-                          style={compactInputStyle}
-                          value={context.opponent.id}
-                        >
-                          {roster.map((participant) => (
-                            <option key={participant.id} value={participant.id}>
-                              {participant.label}
-                            </option>
-                          ))}
-                        </select>
-                      </div>
-                      <div style={rollControlRowStyle}>
-                        <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                          <input
-                            checked={draft.opponentSilent}
-                            disabled={opponentLocked}
-                            onChange={(event) => updateRollDraft(draft.id, { opponentSilent: event.target.checked })}
-                            type="checkbox"
-                          />
-                          Silent
-                        </label>
-                        <span>Use:</span>
-                        <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                          <input
-                            checked={draft.opponentUseGenMod}
-                            disabled={opponentLocked}
-                            onChange={(event) => updateRollDraft(draft.id, { opponentUseGenMod: event.target.checked })}
-                            type="checkbox"
-                          />
-                          Gen
-                        </label>
-                        <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                          <input
-                            checked={draft.opponentUseObSkillMod}
-                            disabled={opponentLocked}
-                            onChange={(event) => updateRollDraft(draft.id, { opponentUseObSkillMod: event.target.checked })}
-                            type="checkbox"
-                          />
-                          OB/Skill
-                        </label>
-                        <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                          <input
-                            checked={draft.opponentUseDbMod}
-                            disabled={opponentLocked}
-                            onChange={(event) => updateRollDraft(draft.id, { opponentUseDbMod: event.target.checked })}
-                            type="checkbox"
-                          />
-                          DB
-                        </label>
-                        <label style={{ alignItems: "center", display: "flex", gap: "0.35rem" }}>
-                          Other
-                          <input
-                            aria-label={`Roleplay roll ${index + 1} opponent Other mod`}
-                            disabled={opponentLocked}
-                            onChange={(event) =>
-                              updateRollDraft(draft.id, {
-                                opponentOtherModInput: event.target.value,
-                                opponentOtherModTouched: true,
-                              })
-                            }
-                            step={1}
-                            style={{ ...compactInputStyle, width: "4.5rem" }}
-                            type="number"
-                            value={context.opponentOtherModInput}
-                          />
-                        </label>
-                      </div>
-                      <div style={rollSkillGridStyle}>
-                        <label style={compactControlStyle}>
-                          <span>Category</span>
-                          <select
-                            aria-label={`Roleplay roll ${index + 1} opponent skill category`}
-                            disabled={opponentLocked}
-                            onChange={(event) => {
-                              const nextCategoryId = event.target.value as RoleplayRollDraft["opponentSkillCategoryId"];
-                              const nextSkillOptions =
-                                nextCategoryId === "all"
-                                  ? context.allOpponentSkillOptions
-                                  : context.allOpponentSkillOptions.filter((skill) => skill.categoryId === nextCategoryId);
-                              updateRollDraft(draft.id, {
-                                opponentSkillCategoryId: nextCategoryId,
-                                opponentRoll: undefined,
-                                opponentOtherModInput: applyUnknownSkillDefaultOtherMod({
-                                  currentValue: draft.opponentOtherModInput,
-                                  selectedSkill: nextSkillOptions[0],
-                                  touched: draft.opponentOtherModTouched,
-                                }),
-                                opponentSkillId: nextSkillOptions[0]?.id ?? "",
-                              });
-                            }}
-                            style={compactInputStyle}
-                            value={draft.opponentSkillCategoryId}
-                          >
-                            <option value="all">All categories</option>
-                            {opponentCategoryOptions.map((category) => (
-                              <option key={category.id} value={category.id}>
-                                {category.label}
-                              </option>
-                            ))}
-                          </select>
-                        </label>
-                        <label style={compactControlStyle}>
-                          <span>Skill</span>
-                          <select
-                            aria-label={`Roleplay roll ${index + 1} opponent skill`}
-                            disabled={opponentLocked || context.opponentSkillOptions.length === 0}
-                            onChange={(event) =>
-                              updateRollDraft(draft.id, {
-                                opponentOtherModInput: applyUnknownSkillDefaultOtherMod({
-                                  currentValue: draft.opponentOtherModInput,
-                                  selectedSkill: context.opponentSkillOptions.find((skill) => skill.id === event.target.value),
-                                  touched: draft.opponentOtherModTouched,
-                                }),
-                                opponentSkillId: event.target.value,
-                              })
-                            }
-                            style={compactSkillInputStyle}
-                            value={context.selectedOpponentSkill?.id ?? ""}
-                          >
-                            <option value="">Choose skill</option>
-                            {context.opponentSkillOptions.length > 0 ? (
-                              context.opponentSkillOptions.map((skill) => (
-                                <option key={skill.id} value={skill.id}>
-                                  {skill.label} ({skill.value ?? 0})
-                                </option>
-                              ))
-                            ) : (
-                              <option value="">No opponent skill</option>
-                            )}
-                          </select>
-                        </label>
-                        <label style={compactControlStyle}>
-                          <span>Support category</span>
-                          <select
-                            aria-label={`Roleplay roll ${index + 1} opponent support skill category`}
-                            disabled={opponentLocked}
-                            onChange={(event) => {
-                              const nextCategoryId = event.target.value as RoleplayRollDraft["opponentSupportSkillCategoryId"];
-                              updateRollDraft(draft.id, {
-                                opponentSupportSkillCategoryId: nextCategoryId,
-                                opponentSupportSkillId: "",
-                              });
-                            }}
-                            style={compactInputStyle}
-                            value={draft.opponentSupportSkillCategoryId}
-                          >
-                            <option value="all">All categories</option>
-                            {opponentSupportCategoryOptions.map((category) => (
-                              <option key={category.id} value={category.id}>
-                                {category.label}
-                              </option>
-                            ))}
-                          </select>
-                        </label>
-                        <label style={compactControlStyle}>
-                          <span>Support</span>
-                          <select
-                            aria-label={`Roleplay roll ${index + 1} opponent support skill`}
-                            disabled={opponentLocked}
-                            onChange={(event) => updateRollDraft(draft.id, { opponentSupportSkillId: event.target.value })}
-                            style={compactSkillInputStyle}
-                            value={context.selectedOpponentSupportSkill?.id ?? ""}
-                          >
-                            <option value="">No support skill</option>
-                            {context.opponentSupportSkillOptions.map((skill) => (
-                              <option key={skill.id} value={skill.id}>
-                                {skill.label} ({skill.value ?? 0})
-                              </option>
-                            ))}
-                          </select>
-                        </label>
-                      </div>
-                      {[context.selectedOpponentSkill, context.selectedOpponentSupportSkill]
-                        .filter((skill): skill is SkillOption => Boolean(skill?.warning))
-                        .map((skill) => (
-                          <div key={skill.id} style={{ color: "#8a5a00", fontSize: "0.85rem" }}>
-                            {skill.label}: {skill.warning}
-                          </div>
-                        ))}
-                      <div style={{ display: "flex", flexWrap: "wrap", gap: "0.5rem" }}>
-                        {opponentStatus ? <span style={{ color: "#5e5a50" }}>{opponentStatus}</span> : null}
-                        <button
-                          disabled={opponentLocked || !context.opponent || !context.selectedOpponentSkill}
-                          onClick={() => void handleAssignSkillRoll(draft, "opponent")}
-                          type="button"
-                        >
-                          Assign
-                        </button>
-                        <button
-                          disabled={opponentLocked || !context.opponent || !context.selectedOpponentSkill}
-                          onClick={() => void handleGmRoll(draft, "opponent")}
-                          type="button"
-                        >
-                          GM Roll
-                        </button>
-                      </div>
-                    </section>
-                    ) : null}
-                  </div>
-                  <RoleplayRollCalculationPanel
-                    actorDifficulty={draft.difficulty === "none" || context.isOpposed ? undefined : draft.difficulty}
-                    actorLabel={`Actor — ${context.participant?.label ?? "Actor"}`}
-                    actorMainPreview={context.preview}
-                    actorPendingLabels={context.preview?.pendingModifierLabels}
-                    actorSupportPreview={context.supportPreview}
-                    comparison={comparison}
-                    opponentMainPreview={context.opponentPreview}
-                    opponentLabel={`Opponent — ${context.opponent?.label ?? "Opponent"}`}
-                    opponentOpen={Boolean(context.opponent && draft.opponentBlockOpen)}
-                    opponentPendingLabels={context.opponentPreview?.pendingModifierLabels}
-                    opponentSupportPreview={context.opponentSupportPreview}
-                  />
-                </section>
-              </div>
+              <RoleplayRollBlock
+                applyUnknownSkillDefaultOtherMod={applyUnknownSkillDefaultOtherMod}
+                comparison={comparison}
+                context={context}
+                draft={draft}
+                index={index}
+                key={draft.id}
+                onActorParticipantChange={handleActorParticipantChange}
+                onAssignSkillRoll={handleAssignSkillRoll}
+                onGmRoll={handleGmRoll}
+                onUpdateRollDraft={updateRollDraft}
+                roster={roster}
+              />
             );
           })}
           <button
@@ -2369,128 +1374,10 @@ export function GmRoleplayingEncounterScreen({
         </div>
       </section>
 
-      <section style={panelStyle}>
-        <h2 style={{ margin: 0 }}>Ranked roll results</h2>
-        {currentRankedRollResults.length > 0 ? (
-          <div style={{ display: "grid", gap: "0.35rem" }}>
-            {currentRankedRollResults.map((entry, index) => {
-              const participantLabel =
-                roster.find((participant) => participant.id === entry.participantId)?.label ??
-                entry.participantId ??
-                "Unknown participant";
+      <RankedRollResultsSection entries={currentRankedRollResults} roster={roster} />
 
-              return (
-                <div key={entry.id}>
-                  {index + 1}. {participantLabel} · {entry.skillLabel ?? "Unknown skill"} ·{" "}
-                  {entry.numericSubtotal == null ? "unresolved" : `total ${entry.numericSubtotal}`}
-                </div>
-              );
-            })}
-          </div>
-        ) : (
-          <div>No GM roll results recorded yet.</div>
-        )}
-      </section>
-
-      <section style={panelStyle}>
-        <h2 style={{ margin: 0 }}>Action log</h2>
-        {roleplayState.actionLog.length > 0 ? (
-          <div style={{ display: "grid", gap: "0.35rem" }}>
-            {roleplayState.actionLog.map((entry) => (
-              <div key={entry.id}>
-                {formatShortDateTime(entry.createdAt)} · {entry.summary}
-                {entry.skillLabel ? ` · ${entry.skillLabel}` : ""}
-                {entry.supportSkillLabel ? ` · Support ${entry.supportSkillLabel}` : ""}
-                {entry.difficulty ? ` · ${formatDifficulty(entry.difficulty)}` : ""}
-                {entry.rollD20 == null ? "" : ` · d20 ${entry.rollD20}`}
-                {entry.openEndedD10s.length > 0 ? ` · d10s ${entry.openEndedD10s.join(", ")}` : ""}
-                {entry.dieResult == null ? "" : ` · die ${entry.dieResult}`}
-                {entry.numericSubtotal == null ? "" : ` · total ${entry.numericSubtotal}`}
-                {entry.opponentSkillLabel ? ` · VERSUS ${entry.opponentParticipantName ?? "Opponent"} ${entry.opponentSkillLabel}` : ""}
-                {entry.opponentRollD20 == null ? "" : ` · opponent d20 ${entry.opponentRollD20}`}
-                {entry.opponentOpenEndedD10s.length > 0 ? ` · opponent d10s ${entry.opponentOpenEndedD10s.join(", ")}` : ""}
-                {entry.opponentDieResult == null ? "" : ` · opponent die ${entry.opponentDieResult}`}
-                {entry.opponentNumericSubtotal == null ? "" : ` · opponent total ${entry.opponentNumericSubtotal}`}
-                {entry.opponentAchievedSuccessLevelLabel ? ` · opponent ${entry.opponentAchievedSuccessLevelLabel}` : ""}
-                {entry.opposedResult ? ` · opposed ${entry.opposedResult}` : ""}
-                {entry.opposedMargin == null ? "" : ` · margin ${entry.opposedMargin}`}
-                {entry.achievedSuccessLevelLabel ? ` · ${entry.achievedSuccessLevelLabel}` : ""}
-                {entry.resultModifier == null ? "" : ` · modifier ${entry.resultModifier >= 0 ? "+" : ""}${entry.resultModifier}`}
-                {entry.success == null ? "" : entry.success ? " · SUCCESS" : " · NOT SUCCESSFUL"}
-                {entry.fumble ? " · FUMBLE" : ""}
-                {entry.partial ? " · Partial" : ""}
-                {entry.useGenMod ? " · Gen mod" : ""}
-                {entry.useObSkillMod ? " · OB/Skill mod" : ""}
-                {entry.useDbMod ? " · DB mod" : ""}
-                {entry.otherMod ? ` · Other ${entry.otherMod > 0 ? "+" : ""}${entry.otherMod}` : ""}
-                {entry.silent ? " · Silent" : ""}
-                {entry.opponentSilent ? " · Opponent silent" : ""}
-                {entry.calculationText ? ` · ${entry.calculationText}` : ""}
-              </div>
-            ))}
-          </div>
-        ) : (
-          <div>No roleplaying actions logged yet.</div>
-        )}
-      </section>
+      <RoleplayActionLogSection entries={roleplayState.actionLog} />
     </section>
-  );
-}
-
-function RoleplayDescriptionRow(input: {
-  description: RoleplayParticipantDescription;
-  onSave: (description: RoleplayParticipantDescription) => void;
-  participant: EncounterParticipant;
-}) {
-  const [name, setName] = useState(input.description.name ?? input.participant.label);
-  const [shortDescription, setShortDescription] = useState(input.description.shortDescription);
-  const [detailedDescription, setDetailedDescription] = useState(input.description.detailedDescription);
-
-  useEffect(() => {
-    setName(input.description.name ?? input.participant.label);
-    setShortDescription(input.description.shortDescription);
-    setDetailedDescription(input.description.detailedDescription);
-  }, [input.description, input.participant.label]);
-
-  return (
-    <details style={{ border: "1px solid #eee8dc", borderRadius: 10, padding: "0.75rem" }}>
-      <summary>
-        <strong>{input.participant.label}</strong>
-      </summary>
-      <div style={{ display: "grid", gap: "0.5rem", marginTop: "0.75rem" }}>
-        <input
-          aria-label={`${input.participant.label} encounter-facing name`}
-          onChange={(event) => setName(event.target.value)}
-          placeholder="Encounter-facing name"
-          value={name}
-        />
-        <input
-          aria-label={`${input.participant.label} short description`}
-          onChange={(event) => setShortDescription(event.target.value)}
-          placeholder="Short description"
-          value={shortDescription}
-        />
-        <textarea
-          aria-label={`${input.participant.label} detailed description`}
-          onChange={(event) => setDetailedDescription(event.target.value)}
-          placeholder="Detailed GM description"
-          rows={3}
-          value={detailedDescription}
-        />
-        <button
-          onClick={() =>
-            input.onSave({
-              detailedDescription,
-              name: name.trim() || input.participant.label,
-              shortDescription,
-            })
-          }
-          type="button"
-        >
-          Save description
-        </button>
-      </div>
-    </details>
   );
 }
