+++ apps/web/app/(app)/encounters/[id]/components/roleplayStyles.ts
@@ -70,6 +70,12 @@ export const rollFieldRowStyle = {
   alignItems: "end",
 } as const;
 
+export const rollSkillGridStyle = {
+  display: "grid",
+  gap: "0.45rem 0.65rem",
+  gridTemplateColumns: "repeat(2, minmax(0, max-content))",
+} as const;
+
 export const rollPreviewStyle = {
   background: "#fbfaf7",
   border: "1px solid #eee8dc",
@@ -91,4 +97,29 @@ export const calculationLineStyle = {
   display: "block",
   overflowX: "auto",
   whiteSpace: "nowrap",
+} as const;
+
+export const playerReadOnlyPanelStyle = {
+  background: "#fbfaf7",
+  border: "1px solid #eee8dc",
+  borderRadius: 8,
+  padding: "0.75rem",
+  whiteSpace: "pre-wrap",
+} as const;
+
+export const playerMetadataTagStyle = {
+  flex: "0 1 auto",
+  fontSize: "0.95rem",
+  fontWeight: 400,
+  minWidth: 0,
+  overflow: "hidden",
+  textOverflow: "ellipsis",
+  whiteSpace: "nowrap",
+} as const;
+
+export const playerRollSkillColumnsStyle = {
+  alignItems: "end",
+  display: "grid",
+  gap: "0.45rem",
+  gridTemplateColumns: "minmax(8rem, 9rem) minmax(10rem, 12rem)",
 } as const;
